Function check_procs::load_procs

fn load_procs(die_on_any_errors: bool) -> RunningProcs

Load currently running procs, and die if there is a surprising error

Normally if this can load any processes it returns what it can find, and prints errors for procs that can't be parsed. But if die_on_any_errors is true it dies if it cannot parse a single process.